requirements.txt
flask
flasgger
python-docx
